+++ crates/teksilo-widgets/src/preview_catalog.rs
@@ -2432,7 +2432,7 @@ impl WidgetCatalog for TreeView<String> {
                         StandardTreeItem::new(lit!(item.clone()))
                             .from_entry(entry)
                             .selected(selected)
-                            .on_toggle_rc(ctx.toggle_callback()),
+                            .on_chevron_toggle_rc(ctx.toggle_callback()),
                     )
                 }),
             ))