+++ examples/widget_catalog/src/tabs/data.rs
@@ -90,7 +90,7 @@ fn tree_view_widget() -> impl Widget + 'static {
             StandardTreeItem::new(lit!(item.clone()))
                 .from_entry(entry)
                 .selected(selected)
-                .on_toggle_rc(ctx.toggle_callback()),
+                .on_chevron_toggle_rc(ctx.toggle_callback()),
         )
     })
     .item_height(28.0)